Move McpHeaderEncoder from Client to Protocol namespace
McpHeaderEncoder is a shared utility used by both client-side (McpHeaderExtractor)
and server-side (StreamableHttpHandler) code. Placing it in the Client namespace
forced server code to reference ModelContextProtocol.Client, creating a misleading
dependency. Move it to ModelContextProtocol.Protocol alongside McpHttpHeaders where
it logically belongs as a transport-level concern.
